requirements.txt
# Core ML and Scientific Computing
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
torch-geometric>=2.3.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Molecular and Chemical Processing
rdkit>=2023.3.1
networkx>=2.6.0

# Protein and NLP Models
transformers>=4.20.0
tokenizers>=0.13.0
datasets>=2.0.0

# Web Framework and API
fastapi>=0.95.0
uvicorn>=0.20.0
streamlit>=1.20.0
pydantic>=1.10.0
python-multipart>=0.0.6

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-decouple>=3.6

# Database and Storage
sqlalchemy>=1.4.0
alembic>=1.8.0
sqlite3

# Visualization and Plotting
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
bokeh>=2.4.0

# Utilities and Development
tqdm>=4.64.0
click>=8.0.0
python-dotenv>=0.19.0
pyyaml>=6.0
requests>=2.28.0

# Testing and Quality
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Jupyter and Development
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.7.0

# Logging and Monitoring
wandb>=0.13.0
tensorboard>=2.10.0
loguru>=0.6.0

# Optional: GPU acceleration (uncomment if using CUDA)
# torch-scatter>=2.1.0
# torch-sparse>=0.6.0
# torch-cluster>=1.6.0

# Optional: Advanced molecular features
# mordred>=1.2.0
# deepchem>=2.6.0